Pass now compiles and runs without segfault. Need to debug annotations.
- Currently, the pass is called after intrinsic cleaner is run, so it
cleans out all of the LLVM annotations.
- What I really need to do is just have the --search=nvm create a flag
so that when KModule is created, it runs this pass and attaches a
NvmFunctionInfo object to the KModule. This is where a lot of the other
passes are getting run anyways. This way, I don't need to change the
intrinsic cleaner or anything.