[DEV-ONLY COMMIT] Stashing before refactor.
- Essentially, I think the symbolic execution version of the heuristic
needs to work at the instruction level (KInstruction) rather than at the
basic block level.
- This is because the state of the values changes at the instruction
level, i.e. the executor executes an instruction, and now we know what object it
now points to rather than speculating about what value it points to. So,
each time we execute an instruction, we need to potentially update the
state of the heuristic. This is confusing at the block level, because
we'd really also need a instruction pointer within a basic block, so why
bother?