Add some useful cmake commands
- klee_add_test_executable compiles a generic executable and extracts
it's bitcode so it can be run in klee. This command cannot take targets
which include "klee.h" or make klee calls.
- klee_add_test_object compiles an object which may include "klee.h" and
make klee calls (like klee_pmem_mark_persistent) and creates a bitcode
that can be run in klee. This is done by compiling the "executable" as a
shared object first (to avoid linking issues with the non-existent
implementations of klee_* calls which are handled by the
SpecialFunctionHandler), then extracting the bitcode.