Actually output errors when programs forget fence
Previously, we would reset the rootCauses array in
PersistentState as soon as the program would clflush/clwb
a cache line. Now we take the same approach as the cache line
tracking, where we have an authoritative rootCauses update list
and a pending one.

Also added test cases that would have caught this.