Add support for nontemporal stores.
Nontemporal stores don't dirty the cache, but we should
still check to make sure that programs put sfences after them.

So when a nontemporal store happens, we dirty the cache line and
immediately flush the same cache line. This leaves a pending
flush in the pending cacheLines update list so that we still
report an error if the program doesn't fence.